Fixing a few small things! :)
The only reason the page wasn't displaying right was because you had the quotes a bit off! Also I went ahead and added closing tags for the div elements and fixed a bit of the indenting for readability! I seriously can't stress enough how amazing you're doing with all of this still though! :) And these things happen even with the best and most experienced web devs, it's so easy to just forget to close a quote!